Added a bunch of deltatime additions, fixed up some of the sensor stuff
The physics doesn't fly everywhere on a mere touch and the sensors now work as they should (provided that you extend the wall sensors just a little so they don't overlap with the ground and air sensors and cause issues), but now the way the controller handles angles is still incredibly busted, and the speed values are way off. Still progress.